对chunklist进行调整，让其适应qInit <--> q000 <-->q025 <--> q050 <-->q075的内存滞后分配/释放

+++ src/main/java/rpc/pool/BuddyChunk.java
@@ -76,8 +76,8 @@ public abstract class BuddyChunk<T> implements PoolChunk {
         return handle;
     }
 
-    public int usage() {
-        return 100 * usedPages / totalPageNum;
+    public double usage() {
+        return 100.0 * usedPages / totalPageNum;
     }
 
     static BuddyChunk newHeapChunk(int pageSize, int maxLevel) {
